Remove old chart controls.
