+++ src/sql/workbench/contrib/charts/browser/chartView.ts
@@ -8,44 +8,17 @@ import 'vs/css!./media/chartView';
 import { IPanelView } from 'sql/base/browser/ui/panel/panel';
 import { Insight } from './insight';
 import QueryRunner from 'sql/workbench/services/query/common/queryRunner';
-import { ChartOptions, IChartOption, ControlType } from './chartOptions';
-import { Extensions, IInsightRegistry, IInsightData } from 'sql/platform/dashboard/browser/insightRegistry';
-import { Registry } from 'vs/platform/registry/common/platform';
+import { IInsightData } from 'sql/platform/dashboard/browser/insightRegistry';
 import * as DOM from 'vs/base/browser/dom';
-import { SelectBox } from 'sql/base/browser/ui/selectBox/selectBox';
-import { IContextViewService } from 'vs/platform/contextview/browser/contextView';
-import { InputBox } from 'vs/base/browser/ui/inputbox/inputBox';
-import { IDisposable, dispose, Disposable } from 'vs/base/common/lifecycle';
-import { attachSelectBoxStyler, attachInputBoxStyler } from 'vs/platform/theme/common/styler';
-import { IThemeService } from 'vs/platform/theme/common/themeService';
+import { Disposable } from 'vs/base/common/lifecycle';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
-import { isUndefinedOrNull } from 'vs/base/common/types';
-import { CreateInsightAction, CopyAction, SaveImageAction, IChartActionContext, ConfigureChartAction } from 'sql/workbench/contrib/charts/browser/actions';
+import { CreateInsightAction, CopyAction, SaveImageAction, ConfigureChartAction } from 'sql/workbench/contrib/charts/browser/actions';
 import { Taskbar } from 'sql/base/browser/ui/taskbar/taskbar';
-import { Checkbox } from 'sql/base/browser/ui/checkbox/checkbox';
-import { IInsightOptions, ChartType } from 'sql/workbench/contrib/charts/common/interfaces';
+import { IInsightOptions, ChartType, DataDirection } from 'sql/workbench/contrib/charts/common/interfaces';
 import { ChartState } from 'sql/workbench/common/editor/query/chartState';
 import * as nls from 'vs/nls';
-import { find } from 'vs/base/common/arrays';
 import { INotificationService } from 'vs/platform/notification/common/notification';
 import { DbCellValue } from 'azdata';
-
-const insightRegistry = Registry.as<IInsightRegistry>(Extensions.InsightContribution);
-
-//Map used to store names and alternative names for chart types.
-//This is mainly used for comparison when options are parsed into the constructor.
-const altNameHash: { [oldName: string]: string } = {
-	'horizontalBar': nls.localize('horizontalBarAltName', "Horizontal Bar"),
-	'bar': nls.localize('barAltName', "Bar"),
-	'line': nls.localize('lineAltName', "Line"),
-	'pie': nls.localize('pieAltName', "Pie"),
-	'scatter': nls.localize('scatterAltName', "Scatter"),
-	'timeSeries': nls.localize('timeSeriesAltName', "Time Series"),
-	'image': nls.localize('imageAltName', "Image"),
-	'count': nls.localize('countAltName', "Count"),
-	'table': nls.localize('tableAltName', "Table"),
-	'doughnut': nls.localize('doughnutAltName', "Doughnut")
-};
 
 export class ChartView extends Disposable implements IPanelView {
 	private insight: Insight;
@@ -61,17 +34,10 @@ export class ChartView extends Disposable implements IPanelView {
 
 	private _state: ChartState;
 
-	private _options: IInsightOptions = {
-		type: ChartType.Bar
-	};
-
+	private _options: IInsightOptions;
 
 	/** parent container */
 	private container: HTMLElement;
-	/** container for the options controls */
-	private optionsControl: HTMLElement;
-	/** container for type specific controls */
-	private typeControls: HTMLElement;
 	/** container for the insight */
 	private insightContainer: HTMLElement;
 	/** container for the action bar */
@@ -79,83 +45,32 @@ export class ChartView extends Disposable implements IPanelView {
 	/** container for the charting (includes insight and options) */
 	private chartingContainer: HTMLElement;
 
-	private optionDisposables: IDisposable[] = [];
 	private optionMap: { [x: string]: { element: HTMLElement; set: (val) => void } } = {};
 
 	constructor(
-		@IContextViewService private _contextViewService: IContextViewService,
-		@IThemeService private _themeService: IThemeService,
 		@IInstantiationService private _instantiationService: IInstantiationService,
 		@INotificationService private readonly _notificationService: INotificationService
 	) {
 		super();
 		this.taskbarContainer = DOM.$('div.taskbar-container');
 		this.taskbar = new Taskbar(this.taskbarContainer);
-		this.optionsControl = DOM.$('div.options-container');
-		const generalControls = DOM.$('div.general-controls');
-		this.optionsControl.appendChild(generalControls);
-		this.typeControls = DOM.$('div.type-controls');
-		this.optionsControl.appendChild(this.typeControls);
 
 		this._createInsightAction = this._instantiationService.createInstance(CreateInsightAction);
 		this._configureChartAction = this._instantiationService.createInstance(ConfigureChartAction, this);
 		this._copyAction = this._instantiationService.createInstance(CopyAction);
 		this._saveAction = this._instantiationService.createInstance(SaveImageAction);
 
-		this.taskbar.setContent([{ action: this._createInsightAction }]);
-
-		const self = this;
-		this._options = new Proxy(this._options, {
-			get: function (target, key) {
-				return target[key];
-			},
-			set: function (target, key, value) {
-				let change = false;
-				if (target[key] !== value) {
-					change = true;
-				}
-
-				target[key] = value;
-				// mirror the change in our state
-				if (self.state) {
-					self.state.options[key] = value;
-				}
-
-				if (change) {
-					self.taskbar.context = <IChartActionContext>{ options: self._options, insight: self.insight ? self.insight.insight : undefined };
-					if (key === 'type') {
-						self.buildOptions();
-					} else {
-						self.verifyOptions();
-					}
-				}
-
-				return true;
-			}
-		}) as IInsightOptions;
-
-
-		ChartOptions.general[0].options = insightRegistry.getAllIds();
-		ChartOptions.general.map(o => {
-			this.createOption(o, generalControls);
-		});
-		this.buildOptions();
+		this._options = {
+			type: ChartType.Bar,
+			dataDirection: DataDirection.Vertical
+		};
 	}
 
 	public clear() {
 
 	}
 
-	/**
-	 * Function used to generate list of alternative names for use with SelectBox
-	 * @param option - the original option names.
-	 */
-	private changeToAltNames(option: string[]): string[] {
-		return option.map(o => altNameHash[o] || o);
-	}
-
 	public dispose() {
-		dispose(this.optionDisposables);
 		super.dispose();
 	}
 
@@ -167,8 +82,8 @@ export class ChartView extends Disposable implements IPanelView {
 			this.container.appendChild(this.taskbarContainer);
 			this.container.appendChild(this.chartingContainer);
 			this.chartingContainer.appendChild(this.insightContainer);
-			this.chartingContainer.appendChild(this.optionsControl);
 			this.insight = new Insight(this.insightContainer, this._options, this._instantiationService);
+			this.updateActionbar();
 		}
 
 		container.appendChild(this.container);
@@ -178,7 +93,6 @@ export class ChartView extends Disposable implements IPanelView {
 		} else {
 			this.queryRunner = this._queryRunner;
 		}
-		this.verifyOptions();
 	}
 
 	public chart(dataId: { batchId: number, resultId: number }) {
@@ -248,45 +162,6 @@ export class ChartView extends Disposable implements IPanelView {
 		}
 	}
 
-	private buildOptions() {
-		// The first element in the disposables list is for the chart type: the master dropdown that controls other option controls.
-		// whiling rebuilding the options we should not dispose it, otherwise it would react to the theme change event
-		if (this.optionDisposables.length > 1) { // this logic needs to be rewritten
-			dispose(this.optionDisposables.slice(1));
-			this.optionDisposables.splice(1);
-		}
-
-		this.optionMap = {
-			'type': this.optionMap['type']
-		};
-		DOM.clearNode(this.typeControls);
-
-		this.updateActionbar();
-		ChartOptions[this._options.type].map(o => {
-			this.createOption(o, this.typeControls);
-		});
-		if (this.insight) {
-			this.insight.options = this._options;
-		}
-		this.verifyOptions();
-	}
-
-	private verifyOptions() {
-		this.updateActionbar();
-		for (let key in this.optionMap) {
-			if (this.optionMap.hasOwnProperty(key)) {
-				let option = find(ChartOptions[this._options.type], e => e.configEntry === key);
-				if (option && option.if) {
-					if (option.if(this._options)) {
-						DOM.show(this.optionMap[key].element);
-					} else {
-						DOM.hide(this.optionMap[key].element);
-					}
-				}
-			}
-		}
-	}
-
 	private updateActionbar() {
 		if (this.insight && this.insight.isCopyable) {
 			this.taskbar.context = { insight: this.insight.insight, options: this._options };
@@ -299,114 +174,6 @@ export class ChartView extends Disposable implements IPanelView {
 		} else {
 			this.taskbar.setContent([{ action: this._createInsightAction }, { action: this._configureChartAction }]);
 		}
-	}
-
-	private createOption(option: IChartOption, container: HTMLElement) {
-		const label = DOM.$('div.option-label');
-		label.innerText = option.label;
-		const optionContainer = DOM.$('div.option-container');
-		const optionInput = DOM.$('div.option-input');
-		optionContainer.appendChild(label);
-		optionContainer.appendChild(optionInput);
-		let setFunc: (val) => void;
-		let value = this.state ? this.state.options[option.configEntry] || option.default : option.default;
-		switch (option.type) {
-			case ControlType.checkbox:
-				let checkbox = new Checkbox(optionInput, {
-					label: '',
-					ariaLabel: option.label,
-					checked: value,
-					onChange: () => {
-						if (this._options[option.configEntry] !== checkbox.checked) {
-							this._options[option.configEntry] = checkbox.checked;
-							if (this.insight) {
-								this.insight.options = this._options;
-							}
-						}
-					}
-				});
-				setFunc = (val: boolean) => {
-					checkbox.checked = val;
-				};
-				break;
-			case ControlType.combo:
-				//pass options into changeAltNames in order for SelectBox to show user-friendly names.
-				let dropdown = new SelectBox(option.displayableOptions || this.changeToAltNames(option.options), undefined, this._contextViewService);
-				dropdown.select(option.options.indexOf(value));
-				dropdown.render(optionInput);
-				dropdown.onDidSelect(e => {
-					if (this._options[option.configEntry] !== option.options[e.index]) {
-						this._options[option.configEntry] = option.options[e.index];
-						if (this.insight) {
-							this.insight.options = this._options;
-						}
-					}
-				});
-				setFunc = (val: string) => {
-					if (!isUndefinedOrNull(val)) {
-						dropdown.select(option.options.indexOf(val));
-					}
-				};
-				this.optionDisposables.push(attachSelectBoxStyler(dropdown, this._themeService));
-				break;
-			case ControlType.input:
-				let input = new InputBox(optionInput, this._contextViewService);
-				input.value = value || '';
-				input.onDidChange(e => {
-					if (this._options[option.configEntry] !== e) {
-						this._options[option.configEntry] = e;
-						if (this.insight) {
-							this.insight.options = this._options;
-						}
-					}
-				});
-				setFunc = (val: string) => {
-					if (!isUndefinedOrNull(val)) {
-						input.value = val;
-					}
-				};
-				this.optionDisposables.push(attachInputBoxStyler(input, this._themeService));
-				break;
-			case ControlType.numberInput:
-				let numberInput = new InputBox(optionInput, this._contextViewService, { type: 'number' });
-				numberInput.value = value || '';
-				numberInput.onDidChange(e => {
-					if (this._options[option.configEntry] !== Number(e)) {
-						this._options[option.configEntry] = Number(e);
-						if (this.insight) {
-							this.insight.options = this._options;
-						}
-					}
-				});
-				setFunc = (val: string) => {
-					if (!isUndefinedOrNull(val)) {
-						numberInput.value = val;
-					}
-				};
-				this.optionDisposables.push(attachInputBoxStyler(numberInput, this._themeService));
-				break;
-			case ControlType.dateInput:
-				let dateInput = new InputBox(optionInput, this._contextViewService, { type: 'datetime-local' });
-				dateInput.value = value || '';
-				dateInput.onDidChange(e => {
-					if (this._options[option.configEntry] !== e) {
-						this._options[option.configEntry] = e;
-						if (this.insight) {
-							this.insight.options = this._options;
-						}
-					}
-				});
-				setFunc = (val: string) => {
-					if (!isUndefinedOrNull(val)) {
-						dateInput.value = val;
-					}
-				};
-				this.optionDisposables.push(attachInputBoxStyler(dateInput, this._themeService));
-				break;
-		}
-		this.optionMap[option.configEntry] = { element: optionContainer, set: setFunc };
-		container.appendChild(optionContainer);
-		this._options[option.configEntry] = value;
 	}
 
 	public set state(val: ChartState) {
