+++ src/sql/workbench/contrib/charts/test/browser/chartView.test.ts
@@ -5,8 +5,6 @@
 
 import * as assert from 'assert';
 import { ChartView } from 'sql/workbench/contrib/charts/browser/chartView';
-import { ContextViewService } from 'vs/platform/contextview/browser/contextViewService';
-import { TestLayoutService } from 'vs/workbench/test/browser/workbenchTestServices';
 import { TestThemeService } from 'vs/platform/theme/test/common/testThemeService';
 import { TestInstantiationService } from 'vs/platform/instantiation/test/common/instantiationServiceMock';
 import { IThemeService } from 'vs/platform/theme/common/themeService';
@@ -25,11 +23,9 @@ suite('Chart View', () => {
 });
 
 function createChartView(): ChartView {
-	const layoutService = new TestLayoutService();
-	const contextViewService = new ContextViewService(layoutService);
 	const themeService = new TestThemeService();
 	const instantiationService = new TestInstantiationService();
 	const notificationService = new TestNotificationService();
 	instantiationService.stub(IThemeService, themeService);
-	return new ChartView(contextViewService, themeService, instantiationService, notificationService);
+	return new ChartView(instantiationService, notificationService);
 }
